requirements.txt
asgiref==3.2.5
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-appconf==1.0.4
django-filter==2.2.0
django-ranged-response==0.2.0
django-select2==7.2.3
django-simple-captcha==0.5.12
djangorestframework==3.11.0
download==0.3.4
fuzzywuzzy==0.18.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.1
idna==2.9
joblib==0.14.1
Markdown==3.2.1
mysqlclient==1.4.6
nltk==3.4.5
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
pdf2image==1.12.1
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytesseract==0.3.3
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tqdm==4.43.0
uritemplate==3.0.1
urllib3==1.25.8
